Move blockmatch to shared location
- Additionally remove outdated blockmatch version
