+++ src/ng_model_gym/usecases/nfru/model/optical_flow/__init__.py
@@ -1,0 +1,23 @@
+# SPDX-FileCopyrightText: <text>Copyright 2026 Arm Limited and/or
+# its affiliates <open-source-office@arm.com></text>
+# SPDX-License-Identifier: Apache-2.0
+
+"""Optical-flow helpers and reference implementations for NFRU."""
+
+from .blockmatch_v32 import BlockMatchV32
+from .blockmatch_v311 import (
+    ArgMinCentered,
+    BlockMatchV311,
+    BlockMatchV311Config,
+    ExtractSearchWindows,
+    upscale_and_dilate_flow,
+)
+
+__all__ = [
+    "ArgMinCentered",
+    "BlockMatchV311",
+    "BlockMatchV311Config",
+    "BlockMatchV32",
+    "ExtractSearchWindows",
+    "upscale_and_dilate_flow",
+]
